fix: :bug: conflict between the formatCurrentWeather function and the variable formatCurrentWeather
The error is caused by a naming conflict between the formatCurrentWeather function and the variable formatCurrentWeather that's declared inside the getFormatteredWeatherData function.  In the getFormatteredWeatherData function, you're using formatCurrentWeather as a variable to store the result of the getWeatherData function after formatting the weather data using the formatCurrentWeather function. However, since you've already declared a function with the same name at the top of the file, the JavaScript interpreter is unable to distinguish between the two and throws an error.  To fix this issue, you should rename the variable formatCurrentWeather to something else, like formattedWeatherData